Refactor App.jsx to extract data into separate modules
Move scene data, markers, and constants out of App.jsx into dedicated files under src/data/ and src/scenes/, and extract shared constants to src/constants.js.
