+++ projects/demo-feature-tour/src/constants.js
@@ -1,0 +1,51 @@
+export const CUSTOM_LAYERS = [
+  'route-outline',
+  'route-line',
+  'nav-puck-layer',
+  'earthquakes-heat',
+  'earthquakes-point',
+  'delivery-zone',
+  'wind-layer',
+  'isochrone-fill-10',
+  'isochrone-fill-20',
+  'isochrone-fill-30',
+  'isochrone-line-10',
+  'isochrone-line-20',
+  'isochrone-line-30',
+  'trail-truck-1',
+  'trail-truck-2',
+  'trail-truck-3',
+  'terrain-trail'
+]
+
+export const CUSTOM_SOURCES = [
+  'route-source',
+  'nav-puck-source',
+  'earthquakes-source',
+  'delivery-zone-source',
+  'mapbox-dem',
+  'raster-array-source',
+  'isochrone-source',
+  'trail-truck-1-source',
+  'trail-truck-2-source',
+  'trail-truck-3-source',
+  'terrain-trail-source'
+]
+
+export const COLOR_THEME_FILTERS = {
+  default: 'none',
+  faded: 'saturate(0.35) brightness(1.09) contrast(0.92)',
+  monochrome: 'grayscale(0.9) contrast(1.1)',
+  ocean: 'hue-rotate(20deg) saturate(1.15) brightness(0.97)',
+  warm: 'sepia(0.28) saturate(1.25) hue-rotate(-8deg)',
+  vivid: 'saturate(1.6) contrast(1.12)'
+}
+
+export const GLOBE_BOUNDS = [
+  [-65, -65],
+  [65, 65]
+]
+
+export const ROTATION_SPEED_FACTOR = 0.02
+
+export const DURATION = 6000
