+++ projects/demo-feature-tour/src/data/navRoute.js
@@ -1,0 +1,306 @@
+export const NAV_ROUTE = [
+  [-122.40512, 37.777406],
+  [-122.404902, 37.777233],
+  [-122.404522, 37.776931],
+  [-122.404634, 37.776842],
+  [-122.404912, 37.776622],
+  [-122.40549, 37.776166],
+  [-122.406217, 37.775591],
+  [-122.406242, 37.775572],
+  [-122.406308, 37.77552],
+  [-122.406412, 37.775603],
+  [-122.406629, 37.775776],
+  [-122.406691, 37.775825],
+  [-122.407091, 37.776145],
+  [-122.407305, 37.776315],
+  [-122.407519, 37.776485],
+  [-122.40756, 37.77652],
+  [-122.407768, 37.776685],
+  [-122.407854, 37.776752],
+  [-122.407955, 37.776833],
+  [-122.408077, 37.776931],
+  [-122.408136, 37.776979],
+  [-122.408187, 37.777019],
+  [-122.40926, 37.777874],
+  [-122.409318, 37.77792],
+  [-122.409388, 37.777976],
+  [-122.40945, 37.778025],
+  [-122.409497, 37.778063],
+  [-122.409945, 37.77842],
+  [-122.410373, 37.778762],
+  [-122.410401, 37.778785],
+  [-122.410562, 37.778913],
+  [-122.410836, 37.779131],
+  [-122.410867, 37.779156],
+  [-122.410936, 37.779211],
+  [-122.411022, 37.77928],
+  [-122.411041, 37.779295],
+  [-122.411239, 37.779453],
+  [-122.41125, 37.779462],
+  [-122.411422, 37.779599],
+  [-122.411484, 37.779648],
+  [-122.411817, 37.779913],
+  [-122.411941, 37.780011],
+  [-122.412039, 37.780091],
+  [-122.412124, 37.780161],
+  [-122.412367, 37.780361],
+  [-122.412409, 37.780395],
+  [-122.412484, 37.780452],
+  [-122.412514, 37.780477],
+  [-122.412528, 37.780488],
+  [-122.412544, 37.780501],
+  [-122.412576, 37.780527],
+  [-122.412649, 37.780584],
+  [-122.412975, 37.780856],
+  [-122.413074, 37.780938],
+  [-122.413149, 37.781001],
+  [-122.413288, 37.780984],
+  [-122.413602, 37.780944],
+  [-122.413693, 37.780933],
+  [-122.413804, 37.780919],
+  [-122.414743, 37.780806],
+  [-122.415245, 37.780738],
+  [-122.415339, 37.780726],
+  [-122.415446, 37.780713],
+  [-122.415652, 37.780687],
+  [-122.415924, 37.780653],
+  [-122.415977, 37.780647],
+  [-122.416161, 37.780623],
+  [-122.416865, 37.780532],
+  [-122.416918, 37.780521],
+  [-122.41699, 37.780499],
+  [-122.417075, 37.780473],
+  [-122.417119, 37.780464],
+  [-122.417285, 37.780443],
+  [-122.417362, 37.780433],
+  [-122.417564, 37.780407],
+  [-122.418082, 37.78034],
+  [-122.418476, 37.780291],
+  [-122.418615, 37.780275],
+  [-122.418748, 37.78026],
+  [-122.418966, 37.780234],
+  [-122.419012, 37.780228],
+  [-122.419534, 37.780161],
+  [-122.419707, 37.780138],
+  [-122.419827, 37.780123],
+  [-122.419844, 37.780121],
+  [-122.420084, 37.780093],
+  [-122.420155, 37.780088],
+  [-122.420156, 37.78009],
+  [-122.420185, 37.780196],
+  [-122.420263, 37.780582],
+  [-122.420335, 37.780945],
+  [-122.42034, 37.780969],
+  [-122.420357, 37.781046],
+  [-122.420377, 37.781134],
+  [-122.420385, 37.781176],
+  [-122.420386, 37.78118],
+  [-122.420453, 37.781514],
+  [-122.420534, 37.781899],
+  [-122.420548, 37.781976],
+  [-122.420565, 37.782059],
+  [-122.420573, 37.782101],
+  [-122.420614, 37.782319],
+  [-122.420633, 37.782507],
+  [-122.420694, 37.782808],
+  [-122.4207, 37.782836],
+  [-122.420721, 37.782911],
+  [-122.420746, 37.782995],
+  [-122.420782, 37.783134],
+  [-122.420829, 37.783379],
+  [-122.420902, 37.783747],
+  [-122.420906, 37.783769],
+  [-122.420921, 37.783841],
+  [-122.420939, 37.783925],
+  [-122.420947, 37.783964],
+  [-122.421018, 37.78431],
+  [-122.421063, 37.784536],
+  [-122.421064, 37.784546],
+  [-122.421075, 37.78465],
+  [-122.421075, 37.784655],
+  [-122.42108, 37.784702],
+  [-122.42109, 37.78478],
+  [-122.421104, 37.784857],
+  [-122.421139, 37.785036],
+  [-122.421181, 37.785245],
+  [-122.42126, 37.785633],
+  [-122.421283, 37.78571],
+  [-122.421308, 37.78579],
+  [-122.421335, 37.785872],
+  [-122.421399, 37.786183],
+  [-122.421466, 37.786511],
+  [-122.421467, 37.786517],
+  [-122.421476, 37.78656],
+  [-122.421493, 37.786636],
+  [-122.421511, 37.786726],
+  [-122.421513, 37.786738],
+  [-122.421514, 37.786744],
+  [-122.421585, 37.787102],
+  [-122.421654, 37.787452],
+  [-122.421663, 37.787496],
+  [-122.421678, 37.787571],
+  [-122.421693, 37.787654],
+  [-122.421695, 37.787667],
+  [-122.421696, 37.78767],
+  [-122.421697, 37.787674],
+  [-122.421698, 37.78768],
+  [-122.421731, 37.787857],
+  [-122.421752, 37.788038],
+  [-122.42183, 37.788425],
+  [-122.421853, 37.788503],
+  [-122.421854, 37.788505],
+  [-122.421878, 37.788587],
+  [-122.421908, 37.788709],
+  [-122.421959, 37.788968],
+  [-122.42204, 37.789356],
+  [-122.422057, 37.789433],
+  [-122.422075, 37.789518],
+  [-122.42208, 37.789541],
+  [-122.422234, 37.790297],
+  [-122.422243, 37.790343],
+  [-122.422244, 37.790346],
+  [-122.42226, 37.790425],
+  [-122.422264, 37.790442],
+  [-122.422275, 37.790496],
+  [-122.422282, 37.790531],
+  [-122.422428, 37.791252],
+  [-122.42244, 37.791316],
+  [-122.422458, 37.791394],
+  [-122.422468, 37.791437],
+  [-122.422488, 37.791524],
+  [-122.422508, 37.791753],
+  [-122.422583, 37.792129],
+  [-122.422602, 37.792195],
+  [-122.422603, 37.792197],
+  [-122.422629, 37.792281],
+  [-122.422648, 37.792332],
+  [-122.422775, 37.792968],
+  [-122.422776, 37.792971],
+  [-122.422777, 37.792975],
+  [-122.422783, 37.79301],
+  [-122.422795, 37.793073],
+  [-122.42281, 37.793147],
+  [-122.422947, 37.793827],
+  [-122.422959, 37.793887],
+  [-122.422974, 37.793953],
+  [-122.422975, 37.793956],
+  [-122.422989, 37.794024],
+  [-122.422993, 37.794064],
+  [-122.422994, 37.794069],
+  [-122.423015, 37.794271],
+  [-122.423106, 37.794724],
+  [-122.423115, 37.794769],
+  [-122.423138, 37.794832],
+  [-122.423167, 37.794908],
+  [-122.423175, 37.794951],
+  [-122.423235, 37.795257],
+  [-122.423293, 37.795619],
+  [-122.423293, 37.795621],
+  [-122.4233, 37.795662],
+  [-122.423316, 37.79576],
+  [-122.423334, 37.795856],
+  [-122.423344, 37.795907],
+  [-122.423346, 37.795919],
+  [-122.423346, 37.795921],
+  [-122.423414, 37.796253],
+  [-122.423421, 37.796291],
+  [-122.423495, 37.796637],
+  [-122.423508, 37.796707],
+  [-122.423523, 37.796796],
+  [-122.423529, 37.796823],
+  [-122.42353, 37.796827],
+  [-122.423625, 37.797278],
+  [-122.423682, 37.79746],
+  [-122.423703, 37.797568],
+  [-122.423718, 37.797638],
+  [-122.423736, 37.797721],
+  [-122.423741, 37.797743],
+  [-122.423741, 37.797745],
+  [-122.423756, 37.797818],
+  [-122.423777, 37.798015],
+  [-122.423853, 37.798392],
+  [-122.423876, 37.798497],
+  [-122.423895, 37.798571],
+  [-122.423925, 37.798656],
+  [-122.423998, 37.798978],
+  [-122.424061, 37.799313],
+  [-122.424075, 37.799386],
+  [-122.424083, 37.79943],
+  [-122.4241, 37.79951],
+  [-122.424113, 37.799579],
+  [-122.424117, 37.799599],
+  [-122.424117, 37.799601],
+  [-122.424117, 37.799603],
+  [-122.424161, 37.799815],
+  [-122.424199, 37.799995],
+  [-122.424202, 37.800009],
+  [-122.424265, 37.800318],
+  [-122.424274, 37.80036],
+  [-122.424293, 37.800434],
+  [-122.424294, 37.800436],
+  [-122.424315, 37.800508],
+  [-122.424341, 37.800622],
+  [-122.424342, 37.800628],
+  [-122.424351, 37.80067],
+  [-122.424393, 37.800871],
+  [-122.424403, 37.800922],
+  [-122.424466, 37.801229],
+  [-122.424467, 37.801232],
+  [-122.424473, 37.801261],
+  [-122.424488, 37.801348],
+  [-122.424635, 37.801346],
+  [-122.424653, 37.801346],
+  [-122.424673, 37.801346],
+  [-122.424766, 37.801338],
+  [-122.424956, 37.801315],
+  [-122.425417, 37.801256],
+  [-122.42568, 37.801222],
+  [-122.426078, 37.801171],
+  [-122.426081, 37.801171],
+  [-122.426111, 37.801167],
+  [-122.426211, 37.801156],
+  [-122.426315, 37.801143],
+  [-122.42632, 37.801142],
+  [-122.426749, 37.801086],
+  [-122.427307, 37.801015],
+  [-122.427739, 37.80096],
+  [-122.427766, 37.800957],
+  [-122.427852, 37.800946],
+  [-122.42796, 37.800932],
+  [-122.428007, 37.800926],
+  [-122.42801, 37.800926],
+  [-122.428248, 37.800893],
+  [-122.428606, 37.800847],
+  [-122.428714, 37.800833],
+  [-122.429385, 37.800748],
+  [-122.429401, 37.800746],
+  [-122.429499, 37.800733],
+  [-122.429588, 37.800722],
+  [-122.429629, 37.800717],
+  [-122.430328, 37.800628],
+  [-122.430805, 37.800567],
+  [-122.430973, 37.800546],
+  [-122.43102, 37.80054],
+  [-122.431047, 37.800536],
+  [-122.431141, 37.800524],
+  [-122.431232, 37.800513],
+  [-122.432001, 37.800415],
+  [-122.432655, 37.800332],
+  [-122.432686, 37.800328],
+  [-122.432786, 37.800315],
+  [-122.432767, 37.800221],
+  [-122.432751, 37.800146],
+  [-122.432691, 37.799849],
+  [-122.433217, 37.799782]
+]
+
+export const NAV_MANEUVERS = [
+  { idx: 2, dir: '↱', instruction: 'Turn right onto Harrison Street' },
+  { idx: 8, dir: '↱', instruction: 'Turn right onto 7th Street' },
+  { idx: 54, dir: '↰', instruction: 'Turn left onto McAllister Street' },
+  { idx: 86, dir: '↱', instruction: 'Turn right onto Van Ness Avenue' },
+  { idx: 248, dir: '↰', instruction: 'Turn left onto Lombard Street' },
+  { idx: 288, dir: '↰', instruction: 'Turn left onto Buchanan Street' },
+  { idx: 291, dir: '↱', instruction: 'Turn right onto Moulton Street' },
+  { idx: 292, dir: '📍', instruction: 'Arrived at destination' }
+]
