+++ projects/demo-feature-tour/src/data/sfMarkers.js
@@ -1,0 +1,65 @@
+export const SF_MARKERS = [
+  {
+    lngLat: [-122.3937, 37.7956],
+    label: 'Ferry Building',
+    emoji: '🏛️',
+    description:
+      'Historic 1898 transit terminal turned artisan marketplace on the Embarcadero waterfront.',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/a/a5/San_Francisco_Ferry_Building_January_2014_panorama.jpg/500px-San_Francisco_Ferry_Building_January_2014_panorama.jpg'
+  },
+  {
+    lngLat: [-122.4089, 37.8087],
+    label: 'Pier 39',
+    emoji: '🎡',
+    description:
+      'Waterfront shopping and entertainment complex, famous for its resident sea lion colony.',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/e/e2/Pier_39%2C_SF%2C_CA%2C_jjron_26.03.2012.jpg/500px-Pier_39%2C_SF%2C_CA%2C_jjron_26.03.2012.jpg'
+  },
+  {
+    lngLat: [-122.4169, 37.8044],
+    label: 'Ghirardelli Square',
+    emoji: '🍫',
+    description:
+      'Former chocolate factory reinvented as a landmark plaza of shops and restaurants.',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/3/3f/Ghiradelli_Square_2.JPG/500px-Ghiradelli_Square_2.JPG'
+  },
+  {
+    lngLat: [-122.405, 37.7849],
+    label: 'Union Square',
+    emoji: '🛍️',
+    description:
+      "San Francisco's premier shopping district, anchored by a public plaza since 1850.",
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/3/32/San_Francisco_-_Union_Square_002.jpg/500px-San_Francisco_-_Union_Square_002.jpg'
+  },
+  {
+    lngLat: [-122.4057, 37.7953],
+    label: 'Chinatown Gate',
+    emoji: '🏮',
+    description:
+      'The ornate Dragon Gate marks the entrance to one of the oldest Chinatowns in North America.',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/8/8e/Dragon_Gate%2C_San_Francisco.jpg/500px-Dragon_Gate%2C_San_Francisco.jpg'
+  },
+  {
+    lngLat: [-122.406, 37.7906],
+    label: 'Transamerica Pyramid',
+    emoji: '🏢',
+    description:
+      'Completed in 1972, this 260-meter tapered spire is the most recognizable landmark on the SF skyline.',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/8/8a/SF_Transamerica_full_CA.jpg/500px-SF_Transamerica_full_CA.jpg'
+  },
+  {
+    lngLat: [-122.3969, 37.8007],
+    label: 'Exploratorium',
+    emoji: '🔬',
+    description:
+      'Museum of science, art, and human perception with over 600 interactive exhibits on Pier 15.',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/d/d9/Exploratorium_Pier_15_%28San_Francisco%29.JPG/500px-Exploratorium_Pier_15_%28San_Francisco%29.JPG'
+  }
+]
