+++ projects/demo-feature-tour/src/data/terrainTrail.js
@@ -1,0 +1,8 @@
+export const TERRAIN_TRAIL_COORDS = [
+  [14.405, 40.814],
+  [14.412, 40.816],
+  [14.418, 40.8178],
+  [14.422, 40.819],
+  [14.4248, 40.8175],
+  [14.426, 40.8211]
+]
